Rearranged Structure. Added new Config (GuildID)
