+++ src/main/java/Commands/Employees/absent.java
@@ -1,4 +1,4 @@
-package Commands;
+package Commands.Employees;
 
 public class absent {
     // TODO
