+++ src/main/java/MainActivity.java
@@ -16,7 +16,7 @@ public class MainActivity implements EventListener {
     public static void main(String[] args) throws LoginException, InterruptedException {
         
         DatabaseHandles db = new DatabaseHandles();
-        TimeThread td = new TimeThread();
+
 
         // Will find the database
         DatabaseParameters.initDB();
@@ -42,6 +42,7 @@ public class MainActivity implements EventListener {
 
         // Calling your name again, remembering all the love you gave to me. Or how i used to be
         commander(jda);
+        TimeThread td = new TimeThread(jda);
 
         td.realtime();
